Working on reducing all state and gate operations to arrays of doubles instead of arrays of complex numbers

+++ src/test/scala/scotty/simulator/DaggerGateSpec.scala
@@ -11,10 +11,10 @@ class DaggerGateSpec extends FlatSpec {
     val gate = (index: Int) => DefGate(Array(
       Array(Complex(1), Complex(0)),
       Array(Complex(0, 1), Complex(0))
-    ), index)
+    ).toDouble, index)
 
     assert(Dagger(gate(0)).matrix(sim).toList.map(_.toList) == List(
       List(Complex(1), Complex(0, -1)),
-      List(Complex(0), Complex(0))))
+      List(Complex(0), Complex(0))).toDouble)
   }
 }
